Refactor settings specs
