+++ src/app/app.module.ts
@@ -106,43 +106,6 @@ import {
 import { VerificationGuard } from './verification-guard';
 
 @NgModule({
-  declarations: [
-    AboutUsComponent,
-    AppComponent,
-    CheckInComponent,
-    CheckInFormComponent,
-    DailyHoursChartComponent,
-    DataDisplayComponent,
-    DataTableComponent,
-    GettingStartedComponent,
-    HeaderComponent,
-    LoginComponent,
-    FindOrganizationComponent,
-    NewVolunteerFormComponent,
-    OrganizationFormComponent,
-    OrganizationConfirmComponent,
-    OrganizationDashboardComponent,
-    PasswordDialogComponent,
-    SettingsPageComponent,
-    SignatureFieldComponent,
-    UserFormComponent,
-    VolunteerMenuComponent,
-    SidenavComponent,
-    ReportsFormComponent,
-    JoinPageComponent,
-    DataCellComponent,
-    EmailDialogComponent,
-    PublicOrganizationDashboardComponent,
-    HomeComponent,
-    ServicesAgreementComponent,
-    ServicesAgreementDialogComponent,
-    LoaderComponent,
-    UserSettingsComponent,
-    OrganizationSettingsComponent,
-    VolunteerSettingsComponent,
-    RolesComponent,
-    ReportsComponent,
-  ],
   imports: [
     AngularFireAuthModule,
     AngularFireModule.initializeApp(environment.firebase),
@@ -188,6 +151,43 @@ import { VerificationGuard } from './verification-guard';
     SignaturePadModule,
     StoreModule.forRoot(reducers),
     StoreRouterConnectingModule,
+  ],
+  declarations: [
+    AboutUsComponent,
+    AppComponent,
+    CheckInComponent,
+    CheckInFormComponent,
+    DailyHoursChartComponent,
+    DataDisplayComponent,
+    DataTableComponent,
+    GettingStartedComponent,
+    HeaderComponent,
+    LoginComponent,
+    FindOrganizationComponent,
+    NewVolunteerFormComponent,
+    OrganizationFormComponent,
+    OrganizationConfirmComponent,
+    OrganizationDashboardComponent,
+    PasswordDialogComponent,
+    SettingsPageComponent,
+    SignatureFieldComponent,
+    UserFormComponent,
+    VolunteerMenuComponent,
+    SidenavComponent,
+    ReportsFormComponent,
+    JoinPageComponent,
+    DataCellComponent,
+    EmailDialogComponent,
+    PublicOrganizationDashboardComponent,
+    HomeComponent,
+    ServicesAgreementComponent,
+    ServicesAgreementDialogComponent,
+    LoaderComponent,
+    UserSettingsComponent,
+    OrganizationSettingsComponent,
+    VolunteerSettingsComponent,
+    RolesComponent,
+    ReportsComponent,
   ],
   providers: [
     AuthService,
