+++ src/app/settings/containers/organization-settings/organization-settings.component.spec.ts
@@ -1,5 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { StoreModule } from '@ngrx/store';
+import { MockComponent } from 'ng2-mock-component';
 
+import * as SettingsActions from '../../../actions/settings.actions';
+import { testOrganizations } from '../../../models';
+import { reducers } from '../../../reducers';
 import { OrganizationSettingsComponent } from './organization-settings.component';
 
 describe('OrganizationSettingsComponent', () => {
@@ -8,9 +13,15 @@ describe('OrganizationSettingsComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ OrganizationSettingsComponent ]
+      imports: [
+        StoreModule.forRoot(reducers),
+      ],
+      declarations: [
+        OrganizationSettingsComponent,
+        MockComponent({ selector: 'app-organization-form', inputs: ['initialOrganization'] }),
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
@@ -21,5 +32,18 @@ describe('OrganizationSettingsComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should handle organizationChanges events by setting organizationChanges', () => {
+    component.onValidOrganization(testOrganizations[0]);
+    expect(component.organizationChanges).toEqual(testOrganizations[0]);
+  });
+
+  it('should handle updateOrganization events by dispatching SettingsActions.UpdateOrganization', () => {
+    spyOn(component.store, 'dispatch');
+    const organization = Object.assign({}, testOrganizations[0], { name: 'Edited' });
+    component.onSubmitOrganization(organization);
+    expect(component.store.dispatch)
+      .toHaveBeenCalledWith(new SettingsActions.UpdateOrganization(organization));
   });
 });
