+++ src/app/settings/containers/reports/reports.component.spec.ts
@@ -1,4 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { StoreModule } from '@ngrx/store';
+import { MockComponent } from 'ng2-mock-component';
+
+import * as SettingsActions from '../../../actions/settings.actions';
+import { testOrganizations, testReports, testVolunteers } from '../../../models';
+import { reducers } from '../../../reducers';
 
 import { ReportsComponent } from './reports.component';
 
@@ -8,9 +14,15 @@ describe('ReportsComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ ReportsComponent ]
+      imports: [
+        StoreModule.forRoot(reducers),
+      ],
+      declarations: [
+        ReportsComponent,
+        MockComponent({ selector: 'app-reports-form' }),
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
@@ -22,4 +34,27 @@ describe('ReportsComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  it('should handle validReport events by setting validReport', () => {
+    component.onValidReport(testReports[0]);
+    expect(component.validReport).toEqual(testReports[0]);
+  });
+
+  it(
+    'should handle generateVisitHistoryReport events by dispatching SettingsActions.GenerateVisitHistoryReport',
+    () => {
+      spyOn(component.store, 'dispatch');
+      component.validReport = testReports[0];
+      component.currentOrganization = testOrganizations[0];
+      component.volunteers = testVolunteers;
+      component.onSubmitReport();
+      expect(component.store.dispatch)
+        .toHaveBeenCalledWith(new SettingsActions.GenerateVisitHistoryReport({
+          startedAt: testReports[0].startedAt,
+          endedAt: testReports[0].endedAt,
+          organizationId: testOrganizations[0].id,
+          volunteers: testVolunteers,
+        }));
+    },
+  );
 });
