+++ src/app/settings/containers/volunteer-settings/volunteer-settings.component.html
@@ -1,6 +1,6 @@
-  <app-data-table
-    [data$]="volunteers$"
-    [columnOptions]="volunteerTableOptions"
-    [showDelete]="true"
-    (deleteItem)="onDeleteVolunteer($event)">
-  </app-data-table>
+<app-data-table
+  [data$]="volunteers$"
+  [columnOptions]="volunteerTableOptions"
+  [showDelete]="true"
+  (deleteItem)="onDeleteVolunteer($event)">
+</app-data-table>
