+++ src/app/settings/containers/volunteer-settings/volunteer-settings.component.spec.ts
@@ -1,4 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { StoreModule } from '@ngrx/store';
+import { MockComponent } from 'ng2-mock-component';
+
+import * as SettingsActions from '../../../actions/settings.actions';
+import { testVolunteers } from '../../../models';
+import { reducers } from '../../../reducers';
 
 import { VolunteerSettingsComponent } from './volunteer-settings.component';
 
@@ -8,9 +14,18 @@ describe('VolunteerSettingsComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ VolunteerSettingsComponent ]
+      imports: [
+        StoreModule.forRoot(reducers),
+      ],
+      declarations: [
+        VolunteerSettingsComponent,
+        MockComponent({
+          selector: 'app-data-table',
+          inputs: ['columnOptions', 'data$', 'showDelete', 'getRowColor'],
+        }),
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
@@ -21,5 +36,13 @@ describe('VolunteerSettingsComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should handle deleteVolunteer events by dispatching SettingsActions.DeleteVolunteer', () => {
+    spyOn(component.store, 'dispatch');
+    const volunteer = testVolunteers[0];
+    component.onDeleteVolunteer(volunteer);
+    expect(component.store.dispatch)
+      .toHaveBeenCalledWith(new SettingsActions.DeleteVolunteer(volunteer));
   });
 });
